Adds basic monitor with simulation
`Monitor` class takes in test data, and, when prompted, sends each packet to the `Method`. It receives packets from the `Method` indicating its prediction, from which the monitor can determine whether the method classified the packet correctly or not.

`Method` is a wrapper for the anomoly model that sniffs all packets classifies them, and sends a packet back to the monitor indicating its prediction. Added a custom header class that indicates the predication and with the original packet.

`Simulation` coordinates packets between the method and monitor instead of having both sniff traffic. When all packets have been sent and classified, prints the results.

Next steps: use an actual dataset rather than a single dummy packet.