Make simulation actually run
In order for the simulation to run with the one dummy packet,
I added a `featurize_scapy_pkt` util function, and I trained
the model at the beginning of the simulation. Note that if you
use a packet file that has a different format than the scapy
packets, then you will need to change the method's conversion
from a packet to a featurized packet that it hands to the model.