Generate dataset from API
Adds the `API` class, which simulates a TCP connection with the autohub API in order to generate packet-level data from the communication. Also adds a command-line parameter to `simulate.py` that specifies a filename for the generated dataset as a pcap file - if the user does not use the command line argument, then the dataset is not saved to a pcap file.